SASS-7607: Upgrade income-tax-additional-information to play 3.0
