+++ build.sbt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 HM Revenue & Customs
+ * Copyright 2024 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-import uk.gov.hmrc.DefaultBuildSettings.integrationTestSettings
-
+import uk.gov.hmrc.DefaultBuildSettings
 val appName = "income-tax-additional-information"
 
-lazy val coverageSettings: Seq[Setting[_]] = {
+lazy val coverageSettings: Seq[Setting[?]] = {
   import scoverage.ScoverageKeys
 
   val excludedPackages = Seq(
@@ -45,20 +44,34 @@ lazy val coverageSettings: Seq[Setting[_]] = {
 }
 
 lazy val microservice = Project(appName, file("."))
-  .enablePlugins(play.sbt.PlayScala, SbtDistributablesPlugin)
-  .settings(
-    // To resolve a bug with version 2.x.x of the scoverage plugin - https://github.com/sbt/sbt/issues/6997
-    libraryDependencySchemes ++= Seq("org.scala-lang.modules" %% "scala-xml" % VersionScheme.Always)
-  )
   .settings(
     majorVersion                     := 0,
     scalaVersion                     := "2.13.12",
-    libraryDependencies              ++= AppDependencies.compile ++ AppDependencies.test
+    libraryDependencies              ++= AppDependencies.compile ++ AppDependencies.test,
   )
-  .configs(IntegrationTest extend Test)
-  .settings(integrationTestSettings(): _*)
+  .configs(Test)
   .settings(resolvers += Resolver.jcenterRepo)
   .settings(PlayKeys.playDefaultPort := 10004)
   .disablePlugins(sbt.plugins.JUnitXmlReportPlugin)
   .disablePlugins(JUnitXmlReportPlugin)
   .settings(coverageSettings: _*)
+
+
+lazy val testSettings: Seq[Def.Setting[?]] = Seq(
+  fork := true,
+  javaOptions ++= Seq("-Dapplication.router=testOnlyDoNotUseInAppConf.Routes"),
+  unmanagedSourceDirectories.withRank(KeyRanks.Invisible) += baseDirectory.value / "test-utils"
+)
+
+lazy val itSettings = DefaultBuildSettings.itSettings() ++ Seq(
+  unmanagedSourceDirectories.withRank(KeyRanks.Invisible) := Seq(
+    baseDirectory.value / "it"
+  )
+)
+
+lazy val it = project
+  .enablePlugins(PlayScala)
+  .dependsOn(microservice % "test->test") // the "test->test" allows reusing test code and test dependencies
+  .settings(testSettings ++ itSettings)
+  .settings(scalaVersion := "2.13.12")
+  .settings(majorVersion := 0)
