+++ it/test/connectors/CreateOrAmendInsurancePoliciesTysConnectorISpec.scala
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package connectors
+package test.connectors
 
 import com.github.tomakehurst.wiremock.http.HttpHeader
 import config.{AppConfig, BackendAppConfig}
-import helpers.WiremockSpec
+import test.helpers.WiremockSpec
 import models._
 import org.scalatestplus.play.PlaySpec
 import play.api.Configuration
@@ -26,22 +26,9 @@ import play.api.http.Status._
 import play.api.libs.json.{JsObject, Json}
 import uk.gov.hmrc.http.{HeaderCarrier, HeaderNames, HttpClient, SessionId}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
-import utils.TaxYearUtils.convertStringTaxYear
+import utils.TaxYearUtils.{convertSpecificTaxYear, specificTaxYear}
 
-class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec {
-
-  lazy val connector: CreateOrAmendInsurancePoliciesConnector = app.injector.instanceOf[CreateOrAmendInsurancePoliciesConnector]
-  implicit val hc: HeaderCarrier = HeaderCarrier()
-  val nino = "nino"
-  val taxYear = 2023
-  val taxYearParameter = convertStringTaxYear(taxYear)
-
-  val url = s"/income-tax/insurance-policies/income/$nino/$taxYearParameter"
-
-  lazy val httpClient: HttpClient = app.injector.instanceOf[HttpClient]
-
-  def appConfig(ifHost: String): AppConfig = new BackendAppConfig(app.injector.instanceOf[Configuration], app.injector.instanceOf[ServicesConfig]) {
-  }
+class CreateOrAmendInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpec {
 
   val model: CreateOrAmendInsurancePoliciesModel = CreateOrAmendInsurancePoliciesModel(
     lifeInsurance = Some(Seq(LifeInsuranceModel(Some("RefNo13254687"), Some("Life"), 123.45, Some(true), Some(4), Some(3), Some(123.45)))),
@@ -53,7 +40,18 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
 
   val modelEmpty: JsObject = Json.obj()
 
-  " PutInsurancePoliciesConnector" should {
+
+  implicit val hc: HeaderCarrier = HeaderCarrier()
+  lazy val connector: CreateOrAmendInsurancePoliciesTysConnector = app.injector.instanceOf[CreateOrAmendInsurancePoliciesTysConnector]
+  lazy val httpClient: HttpClient = app.injector.instanceOf[HttpClient]
+
+  val nino = "nino"
+  val url = s"/income-tax/insurance-policies/income/${convertSpecificTaxYear(specificTaxYear)}/$nino"
+
+  def appConfig(ifHost: String): AppConfig = new BackendAppConfig(app.injector.instanceOf[Configuration], app.injector.instanceOf[ServicesConfig]) {
+  }
+
+  " PutInsurancePoliciesTysConnector" should {
 
     "include internal headers" when {
       val requestBody = Json.toJson(model).toString()
@@ -71,7 +69,7 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
 
         stubPutWithoutResponseBody(url, CREATED, requestBody, headersSentToIf)
 
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Right(expectedResult)
       }
@@ -80,11 +78,11 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
         implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(SessionId("sessionIdValue")))
         val expectedResult = true
 
-        val connector = new CreateOrAmendInsurancePoliciesConnector(httpClient, appConfig(externalHost))
+        val connector = new CreateOrAmendInsurancePoliciesTysConnector(httpClient, appConfig(externalHost))
 
         stubPutWithoutResponseBody(url, CREATED, requestBody, headersSentToIf)
 
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Right(expectedResult)
       }
@@ -97,7 +95,7 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
         stubPutWithoutResponseBody(url, CREATED, Json.toJson(model).toString())
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Right(expectedResult)
       }
@@ -112,7 +110,7 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
 
       stubPutWithResponseBody(url, INTERNAL_SERVER_ERROR, Json.toJson(model).toString(), invalidJson.toString)
       implicit val hc: HeaderCarrier = HeaderCarrier()
-      val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+      val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
       result mustBe Left(expectedResult)
     }
@@ -127,7 +125,7 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
         stubPutWithResponseBody(url, BAD_REQUEST, Json.toJson(model).toString(), responseBody.toString)
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Left(expectedResult)
       }
@@ -148,7 +146,7 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
         stubPutWithResponseBody(url, BAD_REQUEST, Json.toJson(model).toString(), responseBody.toString)
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Left(expectedResult)
       }
@@ -160,10 +158,10 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
           "code" -> "SERVICE_UNAVAILABLE",
           "reason" -> "The service is currently unavailable"
         )
-        stubPutWithResponseBody(url, INTERNAL_SERVER_ERROR, Json.toJson(model).toString(), responseBody.toString)
+        stubPutWithResponseBody(url, SERVICE_UNAVAILABLE, Json.toJson(model).toString(), responseBody.toString)
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Left(expectedResult)
       }
@@ -178,22 +176,23 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
         stubPutWithResponseBody(url, NOT_FOUND, Json.toJson(model).toString(), responseBody.toString)
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Left(expectedResult)
       }
 
       "IF Returns a UNPROCESSABLE_ENTITY" in {
-        val expectedResult = ErrorModel(BAD_REQUEST, ErrorBodyModel("UNPROCESSABLE_ENTITY", "The remote endpoint has indicated that for given income source type, message payload is incorrect."))
+        val expectedResult = ErrorModel(BAD_REQUEST, ErrorBodyModel("UNPROCESSABLE_ENTITY", "The remote endpoint has " +
+          "indicated that for given income source type, message payload is incorrect."))
 
         val responseBody = Json.obj(
           "code" -> "UNPROCESSABLE_ENTITY",
           "reason" -> "The remote endpoint has indicated that for given income source type, message payload is incorrect."
         )
-        stubPutWithResponseBody(url, BAD_REQUEST, Json.toJson(model).toString(), responseBody.toString)
+        stubPutWithResponseBody(url, UNPROCESSABLE_ENTITY, Json.toJson(model).toString(), responseBody.toString)
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Left(expectedResult)
       }
@@ -208,7 +207,7 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
         stubPutWithResponseBody(url, INTERNAL_SERVER_ERROR, Json.toJson(model).toString(), responseBody.toString())
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Left(expectedResult)
       }
@@ -219,7 +218,7 @@ class CreateOrAmendInsurancePoliciesConnectorISpec extends PlaySpec with Wiremoc
         stubPutWithoutResponseBody(url, GONE, Json.toJson(model).toString())
 
         implicit val hc: HeaderCarrier = HeaderCarrier()
-        val result = await(connector.createOrAmendInsurancePolicies(nino, taxYear, model)(hc))
+        val result = await(connector.createOrAmendInsurancePolicies(nino, specificTaxYear, model)(hc))
 
         result mustBe Left(expectedResult)
       }
