+++ it/test/connectors/DeleteInsurancePoliciesConnectorISpec.scala
@@ -14,27 +14,29 @@
  * limitations under the License.
  */
 
-package connectors
+package test.connectors
 
 import com.github.tomakehurst.wiremock.http.HttpHeader
 import config.{AppConfig, BackendAppConfig}
-import helpers.WiremockSpec
+import test.helpers.WiremockSpec
 import models.{ErrorBodyModel, ErrorModel}
 import org.scalatestplus.play.PlaySpec
 import play.api.Configuration
 import play.api.http.Status._
 import uk.gov.hmrc.http.{HeaderCarrier, HeaderNames, HttpClient, SessionId}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
-import utils.TaxYearUtils.convertStringTaxYear
 
-class DeleteOtherEmploymentsIncomeConnectorISpec extends PlaySpec with WiremockSpec {
+class DeleteInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec{
 
-  lazy val connector: DeleteOtherEmploymentsIncomeConnector = app.injector.instanceOf[DeleteOtherEmploymentsIncomeConnector]
+  lazy val connector: DeleteInsurancePoliciesConnector = app.injector.instanceOf[DeleteInsurancePoliciesConnector]
   implicit val hc: HeaderCarrier = HeaderCarrier()
   val nino = "nino"
   val taxYear = 2023
 
-  val ifUrl = s"/income-tax/income/other/employments/${convertStringTaxYear(taxYear)}/$nino"
+  def mkTaxYear(taxYear: Int): String = {
+    s"${taxYear - 1}-${taxYear.toString takeRight 2}"
+  }
+  val ifUrl = s"/income-tax/insurance-policies/income/$nino/${mkTaxYear(taxYear)}"
 
   lazy val httpClient: HttpClient = app.injector.instanceOf[HttpClient]
 
@@ -55,22 +57,22 @@ class DeleteOtherEmploymentsIncomeConnectorISpec extends PlaySpec with WiremockS
 
       "the host for IF is 'Internal'" in {
         implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(SessionId("sessionIdValue")))
-        val connector = new DeleteOtherEmploymentsIncomeConnector(httpClient, appConfig(internalHost))
+        val connector = new DeleteInsurancePoliciesConnector(httpClient, appConfig(internalHost))
 
         stubDeleteWithoutResponseBody(ifUrl, NO_CONTENT, headersSentToIF)
 
-        val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
+        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
 
         result mustBe Right(true)
       }
 
       "the host for IF is 'External'" in {
         implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(SessionId("sessionIdValue")))
-        val connector = new DeleteOtherEmploymentsIncomeConnector(httpClient, appConfig(externalHost))
+        val connector = new DeleteInsurancePoliciesConnector(httpClient, appConfig(externalHost))
 
         stubDeleteWithoutResponseBody(ifUrl, NO_CONTENT, headersSentToIF)
 
-        val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
+        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
 
         result mustBe Right(true)
       }
@@ -80,7 +82,7 @@ class DeleteOtherEmploymentsIncomeConnectorISpec extends PlaySpec with WiremockS
 
       val errorBodyModel = ErrorBodyModel("IF_CODE", "IF_REASON")
 
-      Seq(INTERNAL_SERVER_ERROR, SERVICE_UNAVAILABLE, NOT_FOUND, BAD_REQUEST).foreach { status =>
+      Seq(INTERNAL_SERVER_ERROR, NOT_FOUND, BAD_REQUEST).foreach { status =>
 
         s"If returns $status" in {
           val ifError = ErrorModel(status, errorBodyModel)
@@ -88,7 +90,7 @@ class DeleteOtherEmploymentsIncomeConnectorISpec extends PlaySpec with WiremockS
 
           stubDeleteWithResponseBody(ifUrl, status, ifError.toJson.toString())
 
-          val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
+          val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
 
           result mustBe Left(ifError)
         }
@@ -100,7 +102,19 @@ class DeleteOtherEmploymentsIncomeConnectorISpec extends PlaySpec with WiremockS
 
         stubDeleteWithResponseBody(ifUrl, BAD_GATEWAY, ifError.toJson.toString())
 
-        val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
+        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
+
+        result mustBe Left(ifError)
+      }
+
+
+      "IF returns error code - 503 ServiceUnavailable" in {
+        val ifError = ErrorModel(INTERNAL_SERVER_ERROR, errorBodyModel)
+        implicit val hc: HeaderCarrier = HeaderCarrier()
+
+        stubDeleteWithResponseBody(ifUrl, SERVICE_UNAVAILABLE, ifError.toJson.toString())
+
+        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
 
         result mustBe Left(ifError)
       }
