+++ it/test/connectors/DeleteInsurancePoliciesTysConnectorISpec.scala
@@ -14,29 +14,26 @@
  * limitations under the License.
  */
 
-package connectors
+package test.connectors
 
 import com.github.tomakehurst.wiremock.http.HttpHeader
 import config.{AppConfig, BackendAppConfig}
-import helpers.WiremockSpec
+import test.helpers.WiremockSpec
 import models.{ErrorBodyModel, ErrorModel}
 import org.scalatestplus.play.PlaySpec
 import play.api.Configuration
 import play.api.http.Status._
 import uk.gov.hmrc.http.{HeaderCarrier, HeaderNames, HttpClient, SessionId}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
+import utils.TaxYearUtils.{convertSpecificTaxYear, specificTaxYear}
 
-class DeleteInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec{
+class DeleteInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpec{
 
-  lazy val connector: DeleteInsurancePoliciesConnector = app.injector.instanceOf[DeleteInsurancePoliciesConnector]
+  lazy val connector: DeleteInsurancePoliciesTysConnector = app.injector.instanceOf[DeleteInsurancePoliciesTysConnector]
   implicit val hc: HeaderCarrier = HeaderCarrier()
   val nino = "nino"
-  val taxYear = 2023
 
-  def mkTaxYear(taxYear: Int): String = {
-    s"${taxYear - 1}-${taxYear.toString takeRight 2}"
-  }
-  val ifUrl = s"/income-tax/insurance-policies/income/$nino/${mkTaxYear(taxYear)}"
+  val ifUrl = s"/income-tax/insurance-policies/income/${convertSpecificTaxYear(specificTaxYear)}/$nino"
 
   lazy val httpClient: HttpClient = app.injector.instanceOf[HttpClient]
 
@@ -44,7 +41,7 @@ class DeleteInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec{
     override lazy val ifBaseUrl: String = s"http://$ifHost:$wireMockPort"
   }
 
-  "DeleteInsurancePoliciesConnector " should {
+  "DeleteInsurancePoliciesTysConnector " should {
 
     "include internal headers" when {
       val headersSentToIF = Seq(
@@ -57,22 +54,22 @@ class DeleteInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec{
 
       "the host for IF is 'Internal'" in {
         implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(SessionId("sessionIdValue")))
-        val connector = new DeleteInsurancePoliciesConnector(httpClient, appConfig(internalHost))
+        val connector = new DeleteInsurancePoliciesTysConnector(httpClient, appConfig(internalHost))
 
         stubDeleteWithoutResponseBody(ifUrl, NO_CONTENT, headersSentToIF)
 
-        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
+        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
 
         result mustBe Right(true)
       }
 
       "the host for IF is 'External'" in {
         implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(SessionId("sessionIdValue")))
-        val connector = new DeleteInsurancePoliciesConnector(httpClient, appConfig(externalHost))
+        val connector = new DeleteInsurancePoliciesTysConnector(httpClient, appConfig(externalHost))
 
         stubDeleteWithoutResponseBody(ifUrl, NO_CONTENT, headersSentToIF)
 
-        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
+        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
 
         result mustBe Right(true)
       }
@@ -82,7 +79,7 @@ class DeleteInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec{
 
       val errorBodyModel = ErrorBodyModel("IF_CODE", "IF_REASON")
 
-      Seq(INTERNAL_SERVER_ERROR, NOT_FOUND, BAD_REQUEST).foreach { status =>
+      Seq(BAD_REQUEST).foreach { status =>
 
         s"If returns $status" in {
           val ifError = ErrorModel(status, errorBodyModel)
@@ -90,7 +87,7 @@ class DeleteInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec{
 
           stubDeleteWithResponseBody(ifUrl, status, ifError.toJson.toString())
 
-          val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
+          val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
 
           result mustBe Left(ifError)
         }
@@ -102,19 +99,29 @@ class DeleteInsurancePoliciesConnectorISpec extends PlaySpec with WiremockSpec{
 
         stubDeleteWithResponseBody(ifUrl, BAD_GATEWAY, ifError.toJson.toString())
 
-        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
+        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
 
         result mustBe Left(ifError)
       }
 
-
-      "IF returns error code - 503 ServiceUnavailable" in {
+      "IF returns error code - 503 ServiceUnavailable" in{
         val ifError = ErrorModel(INTERNAL_SERVER_ERROR, errorBodyModel)
         implicit val hc: HeaderCarrier = HeaderCarrier()
 
         stubDeleteWithResponseBody(ifUrl, SERVICE_UNAVAILABLE, ifError.toJson.toString())
 
-        val result = await(connector.deleteInsurancePoliciesData(nino, taxYear)(hc))
+        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
+
+        result mustBe Left(ifError)
+      }
+
+      "IF returns error code - 422 UnprocessableEntity" in {
+        val ifError = ErrorModel(BAD_REQUEST, errorBodyModel)
+        implicit val hc: HeaderCarrier = HeaderCarrier()
+
+        stubDeleteWithResponseBody(ifUrl, UNPROCESSABLE_ENTITY, ifError.toJson.toString())
+
+        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
 
         result mustBe Left(ifError)
       }
