+++ it/test/connectors/DeleteOtherEmploymentsIncomeConnectorISpec.scala
@@ -14,26 +14,27 @@
  * limitations under the License.
  */
 
-package connectors
+package test.connectors
 
 import com.github.tomakehurst.wiremock.http.HttpHeader
 import config.{AppConfig, BackendAppConfig}
-import helpers.WiremockSpec
+import test.helpers.WiremockSpec
 import models.{ErrorBodyModel, ErrorModel}
 import org.scalatestplus.play.PlaySpec
 import play.api.Configuration
 import play.api.http.Status._
 import uk.gov.hmrc.http.{HeaderCarrier, HeaderNames, HttpClient, SessionId}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
-import utils.TaxYearUtils.{convertSpecificTaxYear, specificTaxYear}
+import utils.TaxYearUtils.convertStringTaxYear
 
-class DeleteInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpec{
+class DeleteOtherEmploymentsIncomeConnectorISpec extends PlaySpec with WiremockSpec {
 
-  lazy val connector: DeleteInsurancePoliciesTysConnector = app.injector.instanceOf[DeleteInsurancePoliciesTysConnector]
+  lazy val connector: DeleteOtherEmploymentsIncomeConnector = app.injector.instanceOf[DeleteOtherEmploymentsIncomeConnector]
   implicit val hc: HeaderCarrier = HeaderCarrier()
   val nino = "nino"
+  val taxYear = 2023
 
-  val ifUrl = s"/income-tax/insurance-policies/income/${convertSpecificTaxYear(specificTaxYear)}/$nino"
+  val ifUrl = s"/income-tax/income/other/employments/${convertStringTaxYear(taxYear)}/$nino"
 
   lazy val httpClient: HttpClient = app.injector.instanceOf[HttpClient]
 
@@ -41,7 +42,7 @@ class DeleteInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpe
     override lazy val ifBaseUrl: String = s"http://$ifHost:$wireMockPort"
   }
 
-  "DeleteInsurancePoliciesTysConnector " should {
+  "DeleteInsurancePoliciesConnector " should {
 
     "include internal headers" when {
       val headersSentToIF = Seq(
@@ -54,22 +55,22 @@ class DeleteInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpe
 
       "the host for IF is 'Internal'" in {
         implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(SessionId("sessionIdValue")))
-        val connector = new DeleteInsurancePoliciesTysConnector(httpClient, appConfig(internalHost))
+        val connector = new DeleteOtherEmploymentsIncomeConnector(httpClient, appConfig(internalHost))
 
         stubDeleteWithoutResponseBody(ifUrl, NO_CONTENT, headersSentToIF)
 
-        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
+        val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
 
         result mustBe Right(true)
       }
 
       "the host for IF is 'External'" in {
         implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(SessionId("sessionIdValue")))
-        val connector = new DeleteInsurancePoliciesTysConnector(httpClient, appConfig(externalHost))
+        val connector = new DeleteOtherEmploymentsIncomeConnector(httpClient, appConfig(externalHost))
 
         stubDeleteWithoutResponseBody(ifUrl, NO_CONTENT, headersSentToIF)
 
-        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
+        val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
 
         result mustBe Right(true)
       }
@@ -79,7 +80,7 @@ class DeleteInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpe
 
       val errorBodyModel = ErrorBodyModel("IF_CODE", "IF_REASON")
 
-      Seq(BAD_REQUEST).foreach { status =>
+      Seq(INTERNAL_SERVER_ERROR, SERVICE_UNAVAILABLE, NOT_FOUND, BAD_REQUEST).foreach { status =>
 
         s"If returns $status" in {
           val ifError = ErrorModel(status, errorBodyModel)
@@ -87,7 +88,7 @@ class DeleteInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpe
 
           stubDeleteWithResponseBody(ifUrl, status, ifError.toJson.toString())
 
-          val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
+          val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
 
           result mustBe Left(ifError)
         }
@@ -99,29 +100,7 @@ class DeleteInsurancePoliciesTysConnectorISpec extends PlaySpec with WiremockSpe
 
         stubDeleteWithResponseBody(ifUrl, BAD_GATEWAY, ifError.toJson.toString())
 
-        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
-
-        result mustBe Left(ifError)
-      }
-
-      "IF returns error code - 503 ServiceUnavailable" in{
-        val ifError = ErrorModel(INTERNAL_SERVER_ERROR, errorBodyModel)
-        implicit val hc: HeaderCarrier = HeaderCarrier()
-
-        stubDeleteWithResponseBody(ifUrl, SERVICE_UNAVAILABLE, ifError.toJson.toString())
-
-        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
-
-        result mustBe Left(ifError)
-      }
-
-      "IF returns error code - 422 UnprocessableEntity" in {
-        val ifError = ErrorModel(BAD_REQUEST, errorBodyModel)
-        implicit val hc: HeaderCarrier = HeaderCarrier()
-
-        stubDeleteWithResponseBody(ifUrl, UNPROCESSABLE_ENTITY, ifError.toJson.toString())
-
-        val result = await(connector.deleteInsurancePoliciesData(nino, specificTaxYear)(hc))
+        val result = await(connector.deleteOtherEmploymentsIncomeData(nino, taxYear)(hc))
 
         result mustBe Left(ifError)
       }
