+++ project/plugins.sbt
@@ -18,13 +18,9 @@ resolvers += MavenRepository("HMRC-open-artefacts-maven2", "https://open.artefac
 resolvers += Resolver.url("HMRC-open-artefacts-ivy2", url("https://open.artefacts.tax.service.gov.uk/ivy2"))(Resolver.ivyStylePatterns)
 resolvers += Resolver.typesafeRepo("releases")
 
-// To resolve a bug with version 2.x.x of the scoverage plugin - https://github.com/sbt/sbt/issues/6997
-ThisBuild / libraryDependencySchemes ++= Seq(
-  "org.scala-lang.modules" %% "scala-xml" % VersionScheme.Always
-)
 
 addSbtPlugin("uk.gov.hmrc"       % "sbt-auto-build"         % "3.20.0")
 addSbtPlugin("uk.gov.hmrc"       % "sbt-distributables"     % "2.5.0")
-addSbtPlugin("com.typesafe.play" % "sbt-plugin"             % "2.8.21")
+addSbtPlugin("org.playframework" % "sbt-plugin"             % "3.0.2")
 addSbtPlugin("org.scoverage"     % "sbt-scoverage"          % "2.0.9")
-addSbtPlugin("org.scalastyle"    %% "scalastyle-sbt-plugin" % "1.0.0")
+addSbtPlugin("org.scalastyle"    % "scalastyle-sbt-plugin"  % "1.0.0" exclude("org.scala-lang.modules", "scala-xml_2.12"))
